AndroidCourse_SunshneApp Lesson 1-3
The Sunshine App created when I walk though the Developing Android Apps
in Udacity. This code's knowledge mainly came from the free course
provided online, and the changes were made from a blank project, not the
sample code. Since the course content was based on Lollipo or before,
and after it was created, some of the strcutre and classes changed
(Fragment, Time, Toolbar, etc), so this result is not exactly like the
sample code.