AndroidCourse_SunshineApp_Lesson4A
This is the DBHelper related content in Lesson4A. I didn't include the
testing code. You should be able to find it on Udacity's
SunshineApp_version2 repository.

I keep all the comments from the original start file since they were the
examples of the part I added based on the instructions.